refactor: MovieCard 및 MoviePosterCard 파일 폴더 구조 변경
