+++ src/shared/ui/Card/MovieCard/index.jsx
@@ -1,5 +1,6 @@
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
+import { PlaySVG } from '../../../assets/SVGicons/40pxIcon';
 
 const Wrapper = styled.div`
   display: flex;
@@ -22,7 +23,7 @@ const Poster = styled.img`
     cursor: pointer;
   }
 `;
-const FlexBox = styled.div`
+const TextBox = styled.div`
   width: 100%;
   overflow: auto;
   display: flex;
@@ -41,17 +42,25 @@ const Rating = styled.div`
   font-size: 18px;
   font-weight: 500;
 `;
-
+const FlexBox = styled.div`
+  width: 355px;
+  height: 500px;
+  background: rgb(255, 255, 255, 0.5);
+  border-radius: 5px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
 const MovieCard = (pr) => {
   return (
     <Wrapper>
       <Link to={`/details/${pr.data.id}`}>
-        <Poster src={`https://image.tmdb.org/t/p/w500${pr.data.poster_path}`} />
+        {pr.data.poster_path ? <Poster src={`https://image.tmdb.org/t/p/w500${pr.data.poster_path}`} alt={pr.data.title} /> : <FlexBox>{PlaySVG({ fill: '#aeaeae' })}</FlexBox>}
       </Link>
-      <FlexBox>
+      <TextBox>
         <Title>{pr.data.title}</Title>
         <Rating>평점: {pr.data.vote_average}</Rating>
-      </FlexBox>
+      </TextBox>
     </Wrapper>
   );
 };
