+++ src/shared/ui/Card/MoviePosterCard/index.jsx
